requirements.txt
PyQt5
Pillow
Wand
send2trash
openai
groq